自動実行履歴のElapsedTimeをdoubleに変更、Enum/列コメント整備
- SysHistAutoexec.ElapsedTimeをint→doubleに変更し、1秒未満の処理時間も記録できるようにした
- ToHistoryElapsedSecondsのMath.Ceilingによる秒への切り上げを廃止し、小数を保持するよう修正
- CommentAttribute/OldTableCommentAttrをEnumにも付与できるよう対象を拡張
- 各種Enum値・CodeNameViewにCommentを付与
- Yobi2のコメント誤りを修正

+++ CvBase/BaseDb0Config.cs
@@ -110,7 +110,7 @@ public sealed partial class SysHistAutoexec : BaseDbClass {
 	/// </summary>
 	[ObservableProperty]
 	[Comment("経過時間 (秒)")]
-	public partial int ElapsedTime { get; set; }
+	public partial double ElapsedTime { get; set; }
 	/// <summary>
 	/// 実行結果コード 0:成功、0以外:エラーコード
 	/// </summary>
